backend/pyproject.toml
[tool.poetry]
name = "abilgram"
version = "0.1.0"
description = "telegram inspired chat"
authors = ["Abilhasan Samadov <abil.samedov2016@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.10"
fastapi = "^0.115.4"
python-socketio = {extras = ["asyncio-client"], version = "^5.11.4"}
uvicorn = "^0.32.0"
websockets = "^13.1"
icecream = "^2.1.3"
isort = "^5.13.2"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
dependency-injector = "^4.42.0"
pydantic-settings = "^2.6.1"
alembic = "^1.13.3"
pydantic = "^2.9.2"
supertokens-python = "^0.24.4"
python-jose = "^3.3.0"
email-validator = "^2.2.0"
aiofiles = "^24.1.0"
python-multipart = "^0.0.17"
typesense = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
